+++ decomp/ST.exe/functions/005E7FE0/decomp.c
@@ -116,10 +116,10 @@ LAB_005e829d:
     *(undefined4 *)&this_00->field_0x31 = 0;
     FUN_006e6080(this_00,2,g_startSystem_0081176C->field_0389,(undefined4 *)puVar1);
   }
-  if (-1 < g_startSystem_0081176C->field_02EC) {
+  if (-1 < (int)g_startSystem_0081176C->field_02EC) {
     FUN_006b35d0((int *)g_ddxContext_008075A8,g_startSystem_0081176C->field_02EC);
   }
-  if (-1 < g_startSystem_0081176C->field_0540) {
+  if (-1 < (int)g_startSystem_0081176C->field_0540) {
     FUN_006b35d0((int *)g_ddxContext_008075A8,g_startSystem_0081176C->field_0540);
   }
   g_currentExceptionFrame = local_60.previous;
